pyproject.toml
[project]
name = "finparser"
version = "0.1.0"
description = "Convert financial statement PDFs into structured Excel workbooks"
requires-python = ">=3.11"
dependencies = [
    "google-genai>=1.0",
    "click>=8.1",
    "httpx>=0.27",
    "openpyxl>=3.1",
    "pydantic>=2.0",
    "python-dotenv>=1.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.34",
    "jinja2>=3.1",
    "python-multipart>=0.0.18",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "ruff>=0.8",
]

[project.scripts]
finparser = "finparser.cli:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/finparser"]

[tool.ruff]
src = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]
